Tidy everything up, remove old code
